y2021 d06 p1 - revert to in

+++ y2021/d06/Puzzle.cs
@@ -7,23 +7,23 @@ namespace AdventOfCode.y2021.d06
 {
     public class Puzzle : IPuzzle
     {
-        private readonly List<long> input = new List<long>();
+        private readonly List<int> input = new List<int>();
 
         public Puzzle()
         {
             input = File.ReadAllText(@"y2021\d06\input.puzzle")
                 .Split(',', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)
-                .Select(long.Parse)
+                .Select(int.Parse)
                 .ToList();
         }
         
         public void Part1()
         {
-            var timers = new List<long>(input);
+            var timers = new List<int>(input);
 
             for (int day = 0; day < 80; day++)
             {
-                var newTimers = new List<long>();
+                var newTimers = new List<int>();
 
                 for (int i = 0; i < timers.Count; i++)
                 {
